#12278 add ipaddressfieldserializer to nested serializers
